refactor: use messages-provider instead of custom component
