+++ ui/secret-admin/package.json
@@ -27,10 +27,11 @@
     "babel-jest": "^29.3.1",
     "babel-plugin-transform-import-meta": "^2.2.0",
     "communicator": "https://assets.juno.global.cloud.sap/libs/communicator@2.2.6/package.tgz",
+    "messages-provider": "https://assets.juno.global.cloud.sap/libs/messages-provider@0.1.13/package.tgz",
     "esbuild": "^0.19.12",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
-    "juno-ui-components": "https://assets.juno.global.cloud.sap/libs/juno-ui-components@2.14.0/package.tgz",
+    "juno-ui-components": "https://assets.juno.global.cloud.sap/libs/juno-ui-components@2.14.1/package.tgz",
     "luxon": "^2.3.0",
     "postcss": "^8.4.21",
     "postcss-url": "^10.1.3",
@@ -44,14 +45,14 @@
     "tailwindcss": "^3.3.1",
     "url-state-provider": "https://assets.juno.global.cloud.sap/libs/url-state-provider@1.3.2/package.tgz",
     "util": "^0.12.4",
-    "zustand": "4.3.7"
+    "zustand": "4.5.2"
   },
   "peerDependencies": {
     "luxon": "^2.3.0",
     "prop-types": "^15.8.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "zustand": "4.3.7"
+    "zustand": "4.5.2"
   },
   "scripts": {
     "test": "jest",
