+++ ui/secret-admin/src/App.js
@@ -7,6 +7,7 @@ import React, { useEffect } from "react"
 
 import { AppShell, AppShellProvider } from "juno-ui-components"
 import StoreProvider, { useGlobalsActions } from "./components/StoreProvider"
+import { MessagesProvider } from "messages-provider"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import AppContent from "./AppContent"
 import styles from "./styles.scss"
@@ -35,17 +36,21 @@ const App = (props = {}) => {
   }, [])
 
   return (
+    <MessagesProvider >
     <QueryClientProvider client={queryClient}>
       <AppShell
         pageHeader="Converged Cloud | Secrets"
         embedded={props.embedded === "true" || props.embedded === true}
       >
-        <AsyncWorker />
-      
+        
+          <AsyncWorker />
           <AppContent props={props} />
+        
+        
       
       </AppShell>
     </QueryClientProvider>
+    </MessagesProvider>
   )
 }
