+++ ui/secret-admin/src/components/SecretFormHeader.tsx
@@ -1,0 +1,6 @@
+import { Messages } from "messages-provider"
+const SecretFormHeader: React.FC = () => {
+  return <Messages onDismiss={() => console.log("dismissed!")} />
+}
+
+export default SecretFormHeader
