+++ ui/types/types.ts
@@ -11,6 +11,7 @@ export type Cluster = components["schemas"]["Cluster"]
 export type PluginDefinition = components["schemas"]["PluginDefinition"]
 export type Plugin = components["schemas"]["Plugin"]
 export type PluginPreset = components["schemas"]["PluginPreset"]
+
 export type UpdateClusterInput = {
   clusters: Cluster[]
   action: UpdateObjectAction
@@ -46,6 +47,11 @@ export type KubernetesCondition = {
   status: string
   message?: string
   lastTransitionTime?: string
+}
+
+export type ResultMessage = {
+  message?: string
+  ok: boolean
 }
 
 // some subtypes
